requirements.txt
pandas~=1.4.4
PyYAML~=6.0
numpy~=1.23.3